+++ src/net/atlas/mod.rs
@@ -23,13 +23,13 @@ use regex::Regex;
 use crate::codec::StacksMessageCodec;
 use clarity::codec::StacksMessageCodec as ClarityStacksMessageCodec;
 
+use crate::burnchains::Txid;
+use crate::chainstate::burn::db::sortdb::SortitionDB;
+use crate::chainstate::burn::ConsensusHash;
 use crate::types::chainstate::StacksBlockId;
 use crate::util_lib::boot::boot_code_id;
-use burnchains::Txid;
-use chainstate::burn::db::sortdb::SortitionDB;
-use chainstate::burn::ConsensusHash;
-use util::hash::{to_hex, Hash160, MerkleHashFunc};
-use vm::types::{QualifiedContractIdentifier, SequenceData, TupleData, Value};
+use clarity::vm::types::{QualifiedContractIdentifier, SequenceData, TupleData, Value};
+use stacks_common::util::hash::{to_hex, Hash160, MerkleHashFunc};
 
 use crate::types::chainstate::BlockHeaderHash;